pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "math-agent"
version = "0.1.0"
description = "Math proof agent with unified CLI/API runtimes and optional Lean verification"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.30",
    "google-genai>=1.0",
    "jsonschema>=4.23",
    "openai>=1.30",
    "tomli>=2.0;python_version<'3.12'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
math-agent = "math_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"math_agent.agents" = ["prompts/*.md"]
